server/pyproject.toml
[project]
name = "iot-midterm-server"
version = "0.1.0"
description = "FastAPI server for IoT midterm"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "jinja2>=3.1.4",
    "paho-mqtt>=2.1.0",
    "pydantic-settings>=2.3.4",
    "redis>=5.0.0",
    "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
